+++ SDL/Enemy.h
@@ -20,22 +20,22 @@ public:
 	Enemy();
 	Enemy(SDL_Renderer* renderer, TextureManager* enemyTex);
 
-	void render(SDL_Renderer* renderer, const int camX, const int camY) const;
+	void render(SDL_Renderer* renderer, const float camX, const float camY) const;
 	void spawn();
 	bool isAlive() const;
 
 	void takeDamage(int damage);
 	void kill();
 
-	int getPosX() const;
-	int getPosY() const;
-	int getWidth() const;
-	int getHeight() const;
+	float getPosX() const;
+	float getPosY() const;
+	float getWidth() const;
+	float getHeight() const;
 	Vector2 getEnemyPos() const;
-	void setPosX(int xPos);
-	void setPosY(int yPos);
-	void setVelocityX(int xVel);
-	void setVelocityY(int yVel);
+	void setPosX(float xPos);
+	void setPosY(float yPos);
+	void setVelocityX(float xVel);
+	void setVelocityY(float yVel);
 
 	bool checkCollisionWithEnemy(const Enemy& e) const;
